Updated Room.java
I kept the description for each room in the command and I added attributes of a name and an arraylist of items. A wild idea： Can we write bedroom and lab as instances of room instead of starting a subclass. But if so, how do we deal with the computer?